Init and organize the repository
