+++ nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=C:\\Users\\J392018\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
